feat: 좋아요 이벤트 관련 서비스 코드 몫 처리
