+++ apps/commerce-api/src/test/java/com/loopers/domain/like/LikeSErviceIntegrationTest.java
@@ -11,7 +11,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
-class LikeServiceIntegrationTest {
+class LikeSErviceIntegrationTest {
 
     @Autowired
     private LikeService likeService;
